incremental: add integrateClause method
This method allows to add a clause to a solver that still has an active
trail. By doing so, we can stop search right in the middle, add more
clauses, and continue.

The method will make sure that there are either 2 unassigned literals
in the clause, or otherwise, we correctly jump down to level 0 to
use this clause for propagation.